pushing for v1.02 (multi-platform)
Tons of changes, most of which are really around setting up a Mac build, and most of those won't be reflected in the code that you can see in the scripts. Mostly, this involves setting up ffmpeg with macOS, working out kinds in the code signing and notarization pipeline, etc.